refactor(core)!: rename `mpi.hpp` header to `mpi_space.hpp`
