+++ src/KokkosComm/mpi/handle.hpp
@@ -18,7 +18,7 @@
 
 #include <KokkosComm/fwd.hpp>
 #include <KokkosComm/concepts.hpp>
-#include "mpi.hpp"
+#include "mpi_space.hpp"
 #include "req.hpp"
 
 namespace KokkosComm {
